requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
spacy==3.8.4
pydantic==2.10.4
python-multipart==0.0.20
readability-lxml==0.8.1
feedparser==6.0.11
lxml[html_clean]==5.3.0
httpx==0.28.1
beautifulsoup4==4.12.3
